feat: Next의 Dynamic Routing을 이용하는 폴더구조로 변경
